requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.2
Django==4.0.5
django-allauth==0.45.0
django-celery-results==2.3.0
django-ckeditor==6.4.2
django-cors-headers==3.11.0
django-cuser==2017.3.16
django-debug-toolbar==3.4.0
django-extensions==3.1.3
django-filter==21.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-tinymce==3.4.0
django-user-agents==0.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyngrok==5.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.2.0
redis-server==6.0.9
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sentry-sdk==1.6.0
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.3.0
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
